Refractored adding task logic to addTask reducer
- Switched addTask action payload type to string
